refactor(spring): extract TxExecutor to eliminate duplication across Tx* components
  TxResult, TxTry, and TxValidated shared identical TransactionTemplate
  boilerplate — differing only in the failure predicate. Extract a
  package-private TxExecutor that owns the transactional bookkeeping;
  each component now delegates to a single executor.execute() call with
  Result::isError, Try::isFailure, or Validated::isInvalid respectively.
  No change to the public API or existing tests.
