+++ spring/src/main/java/dmx/fun/spring/TxValidated.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.TransactionDefinition;
 import org.springframework.transaction.support.DefaultTransactionDefinition;
-import org.springframework.transaction.support.TransactionTemplate;
 
 /**
  * Spring component that executes a {@link Validated}-returning action inside a managed
@@ -54,7 +53,7 @@ import org.springframework.transaction.support.TransactionTemplate;
 @Component
 public class TxValidated {
 
-    private final PlatformTransactionManager txManager;
+    private final TxExecutor executor;
 
     /**
      * Creates a {@code TxValidated} backed by the given transaction manager.
@@ -63,7 +62,7 @@ public class TxValidated {
      * @throws NullPointerException if {@code txManager} is {@code null}
      */
     public TxValidated(PlatformTransactionManager txManager) {
-        this.txManager = Objects.requireNonNull(txManager, "txManager");
+        this.executor = new TxExecutor(txManager);
     }
 
     /**
@@ -112,21 +111,7 @@ public class TxValidated {
      * @throws NullPointerException if any argument is {@code null} or if {@code action}
      *                              returns {@code null}
      */
-    @SuppressWarnings("NullAway")
     public <E, A> Validated<E, A> execute(TransactionDefinition def, Supplier<Validated<E, A>> action) {
-        Objects.requireNonNull(def, "def");
-        Objects.requireNonNull(action, "action");
-        var template = new TransactionTemplate(txManager, def);
-        var result = template.execute(status -> {
-            var v = action.get();
-            Objects.requireNonNull(v, "action must not return null");
-            if (v.isInvalid()) {
-                status.setRollbackOnly();
-            }
-            return v;
-        });
-        // template.execute() returns null only if the callback returns null;
-        // the requireNonNull inside the callback above prevents that.
-        return Objects.requireNonNull(result);
+        return executor.execute(def, action, Validated::isInvalid);
     }
 }
